Moves scrum backlogs under namespace scrum.
