+++ app/views/scrum/backlogs/show.json.jbuilder
